--- frontend/src/common/NavBar.tsx
@@ -1,718 +1,0 @@
-import { useContext, memo, useState, type ReactNode } from "react";
-import {
-    AppBar,
-    Toolbar,
-    Typography,
-    Button,
-    Menu,
-    MenuItem,
-    IconButton,
-    Drawer,
-    List,
-    ListItemIcon,
-    ListItemText,
-    Box,
-    useTheme,
-    useMediaQuery,
-    Divider,
-    Stack,
-    Skeleton,
-    ListItemButton,
-    Fade,
-    alpha,
-    type MenuProps,
-} from "@mui/material";
-import type { SxProps, Theme } from "@mui/material/styles";
-import MenuIcon from "@mui/icons-material/Menu";
-import CloseIcon from "@mui/icons-material/Close";
-import KeyboardArrowDownIcon from "@mui/icons-material/KeyboardArrowDown";
-import HomeFilledIcon from "@mui/icons-material/HomeFilled";
-import HomeRoundedIcon from "@mui/icons-material/HomeRounded";
-import EmojiEventsRoundedIcon from "@mui/icons-material/EmojiEventsRounded";
-import AddCircleRoundedIcon from "@mui/icons-material/AddCircleRounded";
-import CasinoRoundedIcon from "@mui/icons-material/CasinoRounded";
-import InsightsRoundedIcon from "@mui/icons-material/InsightsRounded";
-import MenuBookRoundedIcon from "@mui/icons-material/MenuBookRounded";
-import AdminPanelSettingsRoundedIcon from "@mui/icons-material/AdminPanelSettingsRounded";
-import SettingsRoundedIcon from "@mui/icons-material/SettingsRounded";
-import LogoutRoundedIcon from "@mui/icons-material/LogoutRounded";
-import LoginRoundedIcon from "@mui/icons-material/LoginRounded";
-import { AuthContext } from "@/common/AuthContext";
-import { Link, useLocation } from "react-router-dom";
-import { keyframes } from "@mui/system";
-import { palette, timing } from "@/theme/tokens";
-import { gradientTitle } from "@/theme/utils";
-import { pulse } from "@/theme/animations";
-
-// ─── Animations ────────────────────────────────────────────────────────────
-
-// Curtain reveal — unrolls a dropdown from top to bottom.
-const openDown = keyframes`
-    from { opacity: 0; clip-path: inset(0 0 100% 0); }
-    to   { opacity: 1; clip-path: inset(0 0 0% 0);   }
-`;
-
-const EASE = timing.ease;
-
-/**
- * Hover nudge shared by every nav button — a small rightward slide that
- * mirrors the drawer's text animation (theme ListItemButton padding-left),
- * done with translateX so horizontal buttons don't reflow their neighbors.
- */
-const HOVER_TRANSFORM = "translateX(4px)";
-
-/** Spreadable base for any interactive nav button (origin + easing). */
-const interactiveSx: SxProps<Theme> = {
-    transformOrigin: "left center",
-    transition: `all ${timing.normal} ${EASE}`,
-};
-
-/**
- * Frosted top-bar control shared by the Record button and the user-menu
- * button — a translucent bordered pill. The matching hover effect lives in
- * `frostedButtonHover` so callers can extend it (e.g. the Record dot).
- */
-const frostedButtonBase: SxProps<Theme> = {
-    ...interactiveSx,
-    px: 1.75,
-    py: 0.9,
-    fontSize: "0.95rem",
-    fontWeight: 600,
-    borderRadius: 2,
-    border: `1px solid var(--mui-palette-glass-border)`,
-};
-
-/** Shared hover for frosted controls — fills and nudges right. */
-const frostedButtonHover = {
-    background: "var(--mui-palette-glass-fill)",
-    transform: HOVER_TRANSFORM,
-    borderColor: "var(--mui-palette-glass-borderHover)",
-};
-
-// ─── Shared styles ───────────────────────────────────────────────────────────
-
-/** Top-level nav button with an animated underline indicator. */
-const navBtnSx = (active: boolean): SxProps<Theme> => ({
-    ...interactiveSx,
-    position: "relative",
-    minWidth: "auto",
-    px: 1.75,
-    py: 1,
-    mx: 0.25,
-    fontSize: "0.95rem",
-    fontWeight: active ? 700 : 500,
-    color: "inherit",
-    borderRadius: 2,
-    whiteSpace: "nowrap",
-    "&::after": {
-        content: '""',
-        position: "absolute",
-        left: 14,
-        right: 14,
-        bottom: 6,
-        height: "2px",
-        borderRadius: "2px",
-        // Scheme-aware accent (pastel red in dark, deeper rose in light).
-        backgroundColor: "var(--mui-palette-primary-light)",
-        opacity: 0.85,
-        transform: active ? "scaleX(1)" : "scaleX(0)",
-        transformOrigin: "left center",
-        transition: `transform 0.25s ${EASE}`,
-    },
-    "@media (hover: hover)": {
-        "&:hover": {
-            background: "var(--mui-palette-glass-fillSubtle)",
-            transform: HOVER_TRANSFORM,
-            "&::after": { transform: "scaleX(1)" },
-        },
-    },
-});
-
-/** Shared props for every dropdown menu — slides/grows out smoothly. */
-const menuTransition = (
-    origin: "left" | "right",
-): Partial<MenuProps> => ({
-    anchorOrigin: { vertical: "bottom", horizontal: origin },
-    transformOrigin: { vertical: "top", horizontal: origin },
-    TransitionComponent: Fade,
-    transitionDuration: 260,
-    slotProps: {
-        paper: {
-            sx: {
-                mt: 1,
-                overflow: "hidden",
-                transformOrigin: "top center",
-                boxShadow: "var(--mui-palette-appShadow-nav)",
-                animation: `${openDown} 0.26s ${EASE}`,
-                "& .MuiMenuItem-root": {
-                    transition: `background ${timing.fast}, padding-left 0.18s ${EASE}`,
-                    "&:hover": { pl: 2.4 },
-                },
-            },
-        },
-    },
-});
-
-const chevronSx = (open: boolean): SxProps<Theme> => ({
-    fontSize: "1.05rem !important",
-    transition: `transform 0.25s ${EASE}`,
-    transform: open ? "rotate(180deg)" : "rotate(0deg)",
-});
-
-// ─── Logo ────────────────────────────────────────────────────────────────────
-
-const LogoIcon = () => (
-    <Box
-        className="nav-logo"
-        sx={{
-            width: 38,
-            height: 38,
-            borderRadius: 2,
-            background: "var(--mui-palette-glass-fill)",
-            display: "flex",
-            alignItems: "center",
-            justifyContent: "center",
-            mr: 1.25,
-            flexShrink: 0,
-            transition: `transform 0.3s ${EASE}, background 0.3s ${EASE}`,
-            "@media (hover: hover)": {
-                "&:hover": {
-                    transform: "rotate(-8deg) scale(1.08)",
-                    background: "var(--mui-palette-glass-iconHover)",
-                },
-            },
-        }}
-    >
-        <HomeFilledIcon sx={{ color: "text.primary", fontSize: "1.4rem" }} aria-hidden="true" />
-    </Box>
-);
-
-// ─── Mobile drawer building blocks ───────────────────────────────────────────
-
-/**
- * Row styling shared by every drawer destination. Active rows read with the
- * pastel-red accent on three channels at once — tinted fill, a left accent bar,
- * and bolder weight — so the current location is never signalled by color alone.
- */
-const drawerItemSx = (active: boolean, danger = false): SxProps<Theme> => ({
-    position: "relative",
-    minHeight: 48,
-    borderRadius: 2,
-    pl: 2,
-    pr: 1.25,
-    my: 0.25,
-    color: danger ? "error.main" : active ? "primary.light" : "text.primary",
-    ...(active && { bgcolor: (t: Theme) => alpha(t.palette.primary.main, 0.12) }),
-    "& .MuiListItemIcon-root": { color: "inherit", minWidth: 40 },
-    "& .MuiListItemText-primary": { fontSize: "0.95rem", fontWeight: active ? 700 : 500 },
-    "&::before": {
-        content: '""',
-        position: "absolute",
-        left: 4,
-        top: "50%",
-        height: 22,
-        width: 3,
-        borderRadius: 3,
-        bgcolor: "primary.main",
-        transform: `translateY(-50%) scaleX(${active ? 1 : 0})`,
-        transformOrigin: "left center",
-        transition: `transform ${timing.normal} ${EASE}`,
-        "@media (prefers-reduced-motion: reduce)": { transition: "none" },
-    },
-});
-
-const DrawerSectionLabel = ({ children }: { children: ReactNode }) => (
-    <Typography
-        component="div"
-        sx={{
-            px: 2.5,
-            pt: 2,
-            pb: 0.75,
-            fontSize: "0.7rem",
-            fontWeight: 700,
-            letterSpacing: "0.08em",
-            textTransform: "uppercase",
-            color: "text.disabled",
-        }}
-    >
-        {children}
-    </Typography>
-);
-
-interface DrawerItemProps {
-    label: string;
-    icon: ReactNode;
-    to?: string;
-    onClick?: () => void;
-    active?: boolean;
-    danger?: boolean;
-}
-
-/** A single drawer destination: icon + label, with optional Link routing. */
-const DrawerItem = ({ label, icon, to, onClick, active = false, danger = false }: DrawerItemProps) => {
-    const sx = drawerItemSx(active, danger);
-    const body = (
-        <>
-            <ListItemIcon>{icon}</ListItemIcon>
-            <ListItemText primary={label} />
-        </>
-    );
-    return to ? (
-        <ListItemButton
-            component={Link}
-            to={to}
-            onClick={onClick}
-            aria-current={active ? "page" : undefined}
-            sx={sx}
-        >
-            {body}
-        </ListItemButton>
-    ) : (
-        <ListItemButton onClick={onClick} sx={sx}>
-            {body}
-        </ListItemButton>
-    );
-};
-
-
-// ─── Component ─────────────────────────────────────────────────────────────────
-
-const NavBar = () => {
-    const { player, loading, logout } = useContext(AuthContext);
-    const theme = useTheme();
-    const isMobile = useMediaQuery(theme.breakpoints.down("md"));
-    const location = useLocation();
-
-    const [userAnchor, setUserAnchor] = useState<null | HTMLElement>(null);
-    const [mobileDrawerOpen, setMobileDrawerOpen] = useState(false);
-
-    const isActive = (path: string) => location.pathname.startsWith(path);
-    // The combined Games page is served at both /games and /games/current —
-    // highlight the nav entry for either, but not for /games/create or game detail pages.
-    const isGamesActive =
-        location.pathname === "/games" || location.pathname === "/games/current";
-
-    const handleDrawerToggle = () => setMobileDrawerOpen(!mobileDrawerOpen);
-    const closeDrawer = () => setMobileDrawerOpen(false);
-
-    // ─── Mobile drawer ──────────────────────────────────────────────────────
-
-    const mobileDrawer = (
-        <Box
-            role="presentation"
-            sx={{
-                width: "min(86vw, 320px)",
-                height: "100%",
-                display: "flex",
-                flexDirection: "column",
-                bgcolor: "background.paper",
-            }}
-        >
-            {/* Brand header — mirrors the top bar's gradient wordmark on a dark
-                surface, with the close affordance kept clear of the notch. */}
-            <Box
-                sx={{
-                    pt: "calc(env(safe-area-inset-top) + 14px)",
-                    px: 2,
-                    pb: 1.5,
-                    display: "flex",
-                    alignItems: "center",
-                    justifyContent: "space-between",
-                    borderBottom: (t) => `1px solid ${t.palette.divider}`,
-                }}
-            >
-                <Box
-                    component={Link}
-                    to="/"
-                    onClick={closeDrawer}
-                    sx={{
-                        display: "flex",
-                        alignItems: "center",
-                        textDecoration: "none",
-                        minWidth: 0,
-                    }}
-                >
-                    <LogoIcon />
-                    <Typography variant="h6" sx={{ ...gradientTitle, fontSize: "1.2rem" }}>
-                        UBC MJC
-                    </Typography>
-                </Box>
-                <IconButton
-                    onClick={closeDrawer}
-                    aria-label="Close menu"
-                    sx={{ color: "text.secondary", flexShrink: 0 }}
-                >
-                    <CloseIcon />
-                </IconButton>
-            </Box>
-
-            {/* Scrollable destinations; bottom inset keeps the last row clear of
-                the home indicator on gesture-nav phones. */}
-            <Box
-                component="nav"
-                aria-label="Main navigation"
-                sx={{
-                    flex: 1,
-                    overflowY: "auto",
-                    px: 0.5,
-                    pb: "calc(env(safe-area-inset-bottom) + 16px)",
-                }}
-            >
-                <DrawerSectionLabel>Menu</DrawerSectionLabel>
-                <List disablePadding>
-                    <DrawerItem
-                        to="/"
-                        label="Home"
-                        icon={<HomeRoundedIcon />}
-                        active={location.pathname === "/"}
-                        onClick={closeDrawer}
-                    />
-
-                    <DrawerItem
-                        to="/leaderboard"
-                        label="Leaderboard"
-                        icon={<EmojiEventsRoundedIcon />}
-                        active={isActive("/leaderboard")}
-                        onClick={closeDrawer}
-                    />
-
-                    {!loading && player && (
-                        <DrawerItem
-                            to="/games/create"
-                            label="Record Game"
-                            icon={<AddCircleRoundedIcon />}
-                            active={location.pathname.startsWith("/games/create")}
-                            onClick={closeDrawer}
-                        />
-                    )}
-                    <DrawerItem
-                        to="/games"
-                        label="Games"
-                        icon={<CasinoRoundedIcon />}
-                        active={isGamesActive}
-                        onClick={closeDrawer}
-                    />
-                    <DrawerItem
-                        to="/stats/jp"
-                        label="Stats"
-                        icon={<InsightsRoundedIcon />}
-                        active={isActive("/stats")}
-                        onClick={closeDrawer}
-                    />
-                    <DrawerItem
-                        to="/resources"
-                        label="Resources"
-                        icon={<MenuBookRoundedIcon />}
-                        active={isActive("/resources")}
-                        onClick={closeDrawer}
-                    />
-                </List>
-
-                <Divider sx={{ my: 1, mx: 2 }} />
-                <DrawerSectionLabel>Account</DrawerSectionLabel>
-
-                {loading ? (
-                    <Box sx={{ px: 2, py: 1 }}>
-                        <Skeleton variant="rounded" height={44} />
-                    </Box>
-                ) : player ? (
-                    <List disablePadding>
-                        <Box sx={{ px: 2, py: 1, display: "flex", alignItems: "center", gap: 1.25 }}>
-                            <Box
-                                aria-hidden
-                                sx={{
-                                    width: 38,
-                                    height: 38,
-                                    flexShrink: 0,
-                                    borderRadius: "50%",
-                                    display: "grid",
-                                    placeItems: "center",
-                                    fontWeight: 800,
-                                    fontSize: "1rem",
-                                    color: "primary.light",
-                                    bgcolor: (t) => alpha(t.palette.primary.main, 0.16),
-                                }}
-                            >
-                                {player.username.charAt(0).toUpperCase()}
-                            </Box>
-                            <Box sx={{ minWidth: 0 }}>
-                                <Typography noWrap sx={{ fontWeight: 700, fontSize: "0.95rem" }}>
-                                    {player.username}
-                                </Typography>
-                                <Typography sx={{ fontSize: "0.72rem", color: "text.secondary" }}>
-                                    {player.admin ? "Administrator" : "Member"}
-                                </Typography>
-                            </Box>
-                        </Box>
-
-                        {player.admin && (
-                            <DrawerItem
-                                to="/admin"
-                                label="Admin"
-                                icon={<AdminPanelSettingsRoundedIcon />}
-                                active={isActive("/admin")}
-                                onClick={closeDrawer}
-                            />
-                        )}
-                        <DrawerItem
-                            to="/settings"
-                            label="Settings"
-                            icon={<SettingsRoundedIcon />}
-                            active={isActive("/settings")}
-                            onClick={closeDrawer}
-                        />
-                        <Divider sx={{ my: 0.5, mx: 2 }} />
-                        <DrawerItem
-                            label="Log Out"
-                            icon={<LogoutRoundedIcon />}
-                            danger
-                            onClick={() => {
-                                logout();
-                                closeDrawer();
-                            }}
-                        />
-                    </List>
-                ) : (
-                    <List disablePadding>
-                        <DrawerItem
-                            to="/login"
-                            label="Log In"
-                            icon={<LoginRoundedIcon />}
-                            active={isActive("/login")}
-                            onClick={closeDrawer}
-                        />
-                    </List>
-                )}
-            </Box>
-        </Box>
-    );
-
-    return (
-        <AppBar
-            position="sticky"
-            color="primary"
-            sx={{
-                zIndex: (t) => t.zIndex.drawer + 1,
-                transition: `box-shadow 0.3s ${EASE}, background-color 0.3s ${EASE}`,
-                // Translucent frosted bar; text adapts to the scheme so it stays
-                // legible on the dark (near-black) and light (near-white) bars alike.
-                bgcolor: "var(--mui-palette-glass-bar)",
-                color: "text.primary",
-                borderBottom: (t) => `1px solid ${alpha(t.palette.primary.main, 0.25)}`,
-                backdropFilter: "blur(10px)",
-                WebkitBackdropFilter: "blur(10px)",
-                "@supports not ((backdrop-filter: blur(10px)) or (-webkit-backdrop-filter: blur(10px)))": {
-                    bgcolor: "var(--mui-palette-background-default)",
-                },
-            }}
-        >
-            <Toolbar sx={{ minHeight: { xs: 64, md: 76 }, px: { xs: 2, md: 3 } }}>
-                {isMobile && (
-                    <IconButton
-                        color="inherit"
-                        aria-label="open drawer"
-                        edge="start"
-                        onClick={handleDrawerToggle}
-                        sx={{
-                            mr: 1.5,
-                            transition: `transform 0.25s ${EASE}`,
-                            "@media (hover: hover)": {
-                                "&:hover": { transform: "scale(1.15) rotate(6deg)" },
-                            },
-                        }}
-                    >
-                        <MenuIcon />
-                    </IconButton>
-                )}
-
-                <Box
-                    component={Link}
-                    to="/"
-                    sx={{
-                        display: "flex",
-                        alignItems: "center",
-                        color: "inherit",
-                        textDecoration: "none",
-                        mr: { xs: "auto", md: 3 },
-                        transition: `opacity 0.2s ${EASE}`,
-                        "&:hover": { opacity: 0.95 },
-                    }}
-                >
-                    <Typography
-                        variant="h6"
-                        sx={{
-                            ...gradientTitle,
-                            fontSize: { xs: "1.15rem", md: "1.35rem" },
-                            transition: `transform 0.25s ${EASE}, letter-spacing 0.25s ${EASE}`,
-                            transformOrigin: "left center",
-                            "@media (hover: hover)": {
-                                "&:hover": {
-                                    transform: "translateX(2px) scale(1.04)",
-                                    letterSpacing: "0.01em",
-                                },
-                            },
-                        }}
-                    >
-                        UBC MJC
-                    </Typography>
-                </Box>
-
-                {!isMobile && (
-                    <>
-                        <Stack direction="row" spacing={0.25} sx={{ flexGrow: 1 }}>
-                            <Button
-                                color="inherit"
-                                sx={navBtnSx(isActive("/leaderboard"))}
-                                component={Link}
-                                to="/leaderboard"
-                            >
-                                Leaderboard
-                            </Button>
-
-                            <Button
-                                color="inherit"
-                                sx={navBtnSx(isGamesActive)}
-                                component={Link}
-                                to="/games"
-                            >
-                                Games
-                            </Button>
-                            <Button color="inherit" sx={navBtnSx(isActive("/stats"))} component={Link} to="/stats/jp">
-                                Stats
-                            </Button>
-                            <Button color="inherit" sx={navBtnSx(isActive("/resources"))} component={Link} to="/resources">
-                                Resources
-                            </Button>
-                        </Stack>
-
-                        <Stack direction="row" spacing={1} alignItems="center">
-                            {!loading && player?.admin && (
-                                <Button
-                                    color="inherit"
-                                    sx={{ ...navBtnSx(isActive("/admin")), color: "warning.light", fontWeight: 700 }}
-                                    component={Link}
-                                    to="/admin"
-                                >
-                                    Admin
-                                </Button>
-                            )}
-                            {!loading && player && (
-                                <Button
-                                    color="inherit"
-                                    component={Link}
-                                    to="/games/create"
-                                    startIcon={
-                                        <Box
-                                            className="rec-dot"
-                                            sx={{
-                                                width: 9,
-                                                height: 9,
-                                                borderRadius: "50%",
-                                                bgcolor: "primary.main",
-                                                flexShrink: 0,
-                                                opacity: 0,
-                                            }}
-                                        />
-                                    }
-                                    sx={{
-                                        ...frostedButtonBase,
-                                        "@media (hover: hover)": {
-                                            "&:hover": {
-                                                ...frostedButtonHover,
-                                                "& .rec-dot": {
-                                                    animation: `${pulse} 1.6s ease-in-out infinite`,
-                                                    "@media (prefers-reduced-motion: reduce)": {
-                                                        animation: "none",
-                                                    },
-                                                },
-                                            },
-                                        },
-                                    }}
-                                >
-                                    Record
-                                </Button>
-                            )}
-                            {loading ? (
-                                <Skeleton
-                                    variant="rectangular"
-                                    width={100}
-                                    height={38}
-                                    sx={{ borderRadius: 2, bgcolor: "var(--mui-palette-glass-skeleton)" }}
-                                />
-                            ) : player ? (
-                                <>
-                                    <Button
-                                        color="inherit"
-                                        onClick={(e) => setUserAnchor(e.currentTarget)}
-                                        endIcon={<KeyboardArrowDownIcon sx={chevronSx(Boolean(userAnchor))} />}
-                                        sx={{
-                                            ...frostedButtonBase,
-                                            "@media (hover: hover)": {
-                                                "&:hover": frostedButtonHover,
-                                            },
-                                        }}
-                                    >
-                                        {player.username}
-                                    </Button>
-                                    <Menu
-                                        anchorEl={userAnchor}
-                                        open={Boolean(userAnchor)}
-                                        onClose={() => setUserAnchor(null)}
-                                        {...menuTransition("right")}
-                                    >
-                                        <MenuItem component={Link} to="/settings" onClick={() => setUserAnchor(null)}>
-                                            Settings
-                                        </MenuItem>
-                                        <MenuItem
-                                            onClick={() => { logout(); setUserAnchor(null); }}
-                                            sx={{ color: "error.main" }}
-                                        >
-                                            Log Out
-                                        </MenuItem>
-                                    </Menu>
-                                </>
-                            ) : (
-                                <Button
-                                    variant="contained"
-                                    component={Link}
-                                    to="/login"
-                                    sx={{
-                                        ...interactiveSx,
-                                        bgcolor: "var(--mui-palette-glass-cta)",
-                                        color: "text.primary",
-                                        fontWeight: 700,
-                                        px: 2.5,
-                                        py: 1,
-                                        borderRadius: 2,
-                                        border: `1px solid var(--mui-palette-glass-border)`,
-                                        backdropFilter: "blur(4px)",
-                                        "@media (hover: hover)": {
-                                            "&:hover": {
-                                                bgcolor: "var(--mui-palette-glass-fillStrong)",
-                                                transform: HOVER_TRANSFORM,
-                                                boxShadow: "none",
-                                            },
-                                        },
-                                    }}
-                                >
-                                    Log in
-                                </Button>
-                            )}
-                        </Stack>
-                    </>
-                )}
-            </Toolbar>
-
-            <Drawer
-                anchor="left"
-                open={mobileDrawerOpen}
-                onClose={handleDrawerToggle}
-                transitionDuration={300}
-            >
-                {mobileDrawer}
-            </Drawer>
-        </AppBar>
-    );
-};
-
-export default memo(NavBar);
