+++ frontend/src/common/WithNav.tsx
@@ -1,7 +1,7 @@
 import { Suspense } from "react";
 import { Outlet } from "react-router";
 import { Box } from "@mui/material";
-import NavBar from "@/common/NavBar";
+import NavBar from "@/common/navbar";
 import LoadingFallback from "@/common/LoadingFallback";
 
 const WithNav = () => {
