+++ frontend/src/game/LiveGames.tsx
@@ -2,6 +2,7 @@ import { Box, Card, Skeleton, Alert } from "@mui/material";
 import { gameRoundString } from "./common/constants";
 import GameSummaryCard, { gameSummaryGrid } from "./common/GameSummaryCard";
 import { useLiveGames } from "@/hooks/GameHooks";
+import { formatRelativeTime } from "@/common/Utils";
 import type { GameVariant } from "@/types";
 
 const LiveGamesSkeleton = () => (
@@ -24,25 +25,6 @@ const LiveGamesSkeleton = () => (
 
 export const LiveGamesSection = ({ gameVariant }: { gameVariant: GameVariant }) => {
     const { isPending, error, data: liveGames } = useLiveGames(gameVariant);
-
-    const formatDate = (dateString: string) => {
-        const date = new Date(dateString);
-        const now = new Date();
-        const diffMs = now.getTime() - date.getTime();
-        const diffMins = Math.floor(diffMs / 60000);
-        const diffHours = Math.floor(diffMins / 60);
-
-        if (diffMins < 1) return "Just now";
-        if (diffMins < 60) return `${diffMins}m ago`;
-        if (diffHours < 24) return `${diffHours}h ago`;
-
-        return date.toLocaleDateString("en-US", {
-            month: "short",
-            day: "numeric",
-            hour: "2-digit",
-            minute: "2-digit",
-        });
-    };
 
     if (isPending) {
         return <LiveGamesSkeleton />;
@@ -79,7 +61,7 @@ export const LiveGamesSection = ({ gameVariant }: { gameVariant: GameVariant }) 
                                 game={game}
                                 gameVariant={gameVariant}
                                 chipLabel={gameRoundString(game, gameVariant)}
-                                timeText={formatDate(game.createdAt)}
+                                timeText={formatRelativeTime(game.createdAt)}
                             />
                         ))}
                     </Box>
